Wrangled Data
Wrangled data into desired format (factor conversion, selected wanted variables, accounted for N/A's) and calculated summary statistics. A scatter plot was then made to visualize Age and played_hours, with subscription status being the color of the points. The data is now ready to be split in order to create a classification model.